enc: support global scaling
This adds encoder support to configure and use global scaling.

The configuration option codingScale is a scale factor used to scale the
input point cloud for geometry and attribute coding.

The global scaling value is a scale factor applied at the decoder during
output.  It is determined by the encoder as the ratio of the
sequenceScale to the codingScale.

The codingScale must be less than or equal to the sequenceScale.  The
encoder will adjust the codingScale value if this is not the case.