hls/m56959: add frame_ctr_lsb to frame boundary marker
This adds the ability to identify the frame that the frame boundary
marker terminates.

The decoder state machine is updated to use the frame counter to
detect marked boundaries using the frame ctr value.  The decoder
no longer flushes the output at the moment the marker is received,
rather it waits until the boundary is indicated by a change in
frame_ctr_lsb.