hls/m56959: decouple sequence bounding box origin and size
This permits signalling a sequence origin without specifying a bounding
box size.  The sequence bounding box presence flag is removed.  Instead,
the origin / size bits field is used to indicate the presence of an
origin or bounding box size.

The coding of the bounding box size is corrected to be in minus1 form.