attr: defer resolution of indirect point indexes until after nn search
During LoD construction, the LoD refinement points array (indexes)
contains indexes to points in the sorted Morton order array.  To
avoid double indirection to lookup a point coordinate / attribute
(ie, refinement point index -> Morton index -> point index), the
indexes array is updated to contain point indexes.

This commit defers the update until after the nearest neighbour
search, rather than being performed as a side-effect of the search.

This change is necessary to update the intra search to use a use
previous points (in left-to-right coding order).  It also removes
the need to restore the original index values in the case of
scalable coding changing the reference LoD for the inter search.