attr: remove reversal of lod point index and predictor arrays
During the nearest neighbour search, the point predictor array is
written in reverse order to the search order.  After the search is
complete, the lod refinement point index array is reversed to match the
predictor array.  Subsequent processing then iterates over these arrays
in left-to-right order.

The reversal hides the actual order that points are processed within
an lod.  This commit removes the reversal and explicitly shows where
traversal is counter to the left-to-right order by use of inReverse().