attr/m58313: code attributes in left-to-right order
This reverses the coding order of attributes such that all transform
and coding operations are performed in the same order that points are
output by the geometry coder.

NB: in order to satisfy intra-prediction, intra-LoD predictors are
searched on the 'left' of the current point (i.e., the previous points
in coding order), rather than to the 'right'.