predgeom/m60150+m60236: configurable pred list size
Prediction list size can be configured by the encoder for the decoder.
A new GPS attribute is used (predgeom_max_pred_index) to specify to
prediction list size to the decoder.
The maximum allowed prediction list size is 8.

Two options are added to the encoder:
- predGeomMaxPredIdx, specifying the value of predgeom_max_pred_index,
- predGeomMaxPredIdxTested, specifying the maximum number of prectictors
to be tested by the encoder.