attr: m57324 - improved neighbour search for prediction
Currently up to 3 neighbours may be chosen as predictor for attributes.
It is proposed to replace the 3rd candidate using another predictor if
certain conditions apply - mainly to ensure that neighbours don't fall in
the same region.

There is currently a crash/mismatch for when attribute inter prediction
is enabled; hence tool is disabled in this case until bug is resolved..