misc: Separate intra and inter cfg files
The configuration files for inter prediction are separated from
those of intra. The yaml files are placed in separate folder
cfg/inter/ and the gen-cfg.sh script is modified to generate the
inter prediction cfg files.

Generating the configuration files will produce
the inter cfg files in folders with suffixes "-inter". For example,
cfg files for octree-predlift configuration with inter prediction
will be generated in octree-predlift-inter/. The folder names of
intra prediction will be as before. E.g., cfg files for octree-predlift
with intra prediction will be generated in octree-predlift/ (as before).