attr/fix: m61223 - fixes and improvements to distribution-based neighbour search
Aspect 3: The distribution-based neighbour search method is used to
replace the third predictor in the predictor list with a predictor
that is farther from the current point but results in a better distribution
of the predictor points. Current syntax was identified to have some
bugs and other shortcomings.

Proposal includes the following changes:
- Add a strict opposite check for P1 and P2, where PN is the N-th neighbour
- Use single threshold to check whether directions of P3..Pn
  are strictly opposite to direction of P0 or P1.
- When strict opposite directions are not present, do not replace
  if P1 and P2 are loosely opposite to each other.
- When P1 and P2 are equal, replace P2 (if not already so) with
  a loosely opposite direction to P0/P1.
- The proposed implementation already takes care of the minor software bugs