octree: m60697 - common OBUF class for trisoup/new octree
Dynamic OBUF is applied for the new octree framework.
Trisoup framework also has dynamic OBUF but there are
minor differences between the dynamic OBUF schemes of
octree and trisoup.

This modification aligns the dynamic OBUF scheme used
in octree and trisoup so that only one dymanic OBUF
mechanism is used.