trisoup/m60017: adaptive halo for ray tracing
An optional adaptive halo is added for improving trisoup's ray tracing,
when halo is used.

By default, and for common test condtitions, adaptive halo is always
activated when trisoup halo is enabled. But, it has been demonstrated
that the control of adaptive halo could improve reconstruction on some
rate points and/or sequences.

This commit also modify configuration files to disable trisoup halo on
ulb_unicoorn_vox13 where it fails to provide quality improvement.