trisoup/m61436: fixing single bit vertex position coding
As exposed in m61436's issue 1:
When a single bit was used for representing the vertex position, two bits were
signaled instead of one. It was making the decoded vertices missmatching the
encoded ones and coudld make the decoder crashing.

This commit fixes issue 1 by not encoding/decoding the extra bit when
single bit is used for vertex position.