trisoup/m61436: fixing negative bitshift
As exposed in m61436's issue 2:
When a single bit was used for representing the vertex position,
'nbitsVertices – 2' becomes negative. As it is used as exponent of a bitwise
shift operation during context determination, the result is undefined
according to C++.

This commit fixes issue 2 by clipping the bitwise shift operation to
zero.