fix: m61223 - signalling of flag for modified neighbour search
Aspect 2: The distribution-based neighbour search method is only
applicable when the number of direct predictors for attribute prediction
is at least 3. Currently, the flag enabling the distribution-based
neighbour search is signalled irrespective of the number of direct
predictors.

Proposal is to avoid signalling the flag (infer to 0) that enables
the distribution-based neighbour serach when the number of direct
predictors is less than 3.