geom/inter: Enable encoder derivation of global motion thresholds
The global motion thresholds are used to determine the points to
which motion compensation is to be applied under octree geometry
(when LPU type is 0) or under predictive geometry. Currently the
global motion thresholds are input using the global motion parameters
file.

This solution adds an option to derive the global motion thresholds
at the encoder. The threshold derivation is enabled automatically
when internal GM derivation is used.