attr/inter: m60200 - support slice-level RDO selection of inter/intra pred for attr
This is encoder-only method to enable RDO selectio of inter/intra prediciton
of attributes at the slice level. When enabled, each slice is coded with both
inter prediction and intra prediction separately, and a cost is calculated
based on the rate-distortion. The mode with the lower cost is chosen to code
the attribute for the slice.