predgeom/inter: m61586 - Additional inter prediction candidates
Inter prediction for predictive geometry currently uses two
inter prediction candidates. The two candidates are obtained
using a combination of the zero-compensated reference frame
and global-motion-compensated reference frame.

In this proposal, two additional inter prediction candidates are
added that are obtained from the global-motion compensated
reference; this is used for the radius. The azimuth predictor is
obtained from the parent node. The proposal also includes
changes to the context derivation for the various predgeom syntax
elements.

An additional change to the CTC is proposed which is included in
a separate commit. This changes the first two inter prediction
candidates to use zero motion compensated reference frame.