attr/inter: m61886 - enable inter prediction for RAHT attribute coding
Currently, attribute inter prediction is only enabled for the
predicting and lifting schemes.

This proposal enables inter prediction of RAHT attribute coding. The
DC coefficient for the root node is predicted from reference frame.
The AC coefficient predictor is obtained by averaging one or more
points in the reference frame that have a similar position as the
current node. Inter prediction is only applied to the first five layers
of RAHT coefficients.