predgeom/inter: m62218 - disable addn candidates when GM is disabled
The propsal m61586 add two new inter prediction candidates that
are derived from the global motion compensated reference frame.
The indices to the inter prediction candidate use these new
candidates even when global motion is disabled.

In this change, the two additional candidates are disallowed
when global motion is disabled.