misc: m62100 - add informative cognata sequences
Four sequences obtained from the Cognata simulator are added
in the configuration files for informative purpose. These sequences
are not currently part of the CTC.

The Cognata simulator was modelled to output point clouds as
though generated by a spinning LIDAR sensor with 128 lasers. The
sequences are low noise and contain positions with mm-accuracy.
The sequences were simulated in a city environment.