fix: issue # 49 - implementation of m58315 incomplete
m58315 proposed to subsume fixed QP offset values into the
quantization scaling operation. An earlier commit for the proposal
only applied the change to colours, but not reflectance.
Moreover, in the quantization stage at encoder, a scaling operation
was performed at an earlierstep where the weights are calculated
rather than at the quantization step. This results in loss of
coding efficiency.

The implementaion is completed for reflectances in this commit
and the bug in the quantization step is fixed.