predgeom: m63252 - remove parsing dependence in predgeom
Currently, in predictive geometry, the azimuth residual has a dependence
on boundPhi, which in turn is dependent on the reconstructed radius
of the preceding points. This introduces dependence of parsing on the
decoding/reconstruction of points.

The proposal removes this dependence. The parsing of azimuth residual
is no longer dependent on boundPhi.