fix/hls: m62760 - signalling enableAttrInterPred in abh
Currently, a flag is signalled in the attribute brick header
that may be used to disable attribute inter prediction. This
flag is signalled irrespective of whether attribute inter
prediction is enabled.

The proposal negates the meaning of this flag - as in the
flag is used to enable attribute inter prediction at the brick
header. Moreover, it is only signalled when attribute inter
prediction is enabled at the APS.